Player Controller
This controller script handles movement and collision detection. We move the player through its transform by adding a Vector3 we call playerVelocity. Currently, there is an issue with the player being able to press a direction on the "Horizontal" axis and jumping at the same time. The player does not jump and doesn't add to its playerVelovity.